requirements.txt
# CLIP (not on PyPI — install separately)
# pip install git+https://github.com/openai/CLIP.git

blinker==1.9.0
click==8.3.2
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
fsspec==2026.3.0
ftfy==6.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
packaging==26.0
pillow==12.2.0
regex==2026.4.4
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing_extensions==4.15.0
wcwidth==0.6.0
Werkzeug==3.1.8
